Split index.ts into app & index
